requirements.txt
flask>=3.0.0
anthropic>=0.34.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
qrcode>=7.4.2
pillow>=10.0.0
pywebview>=5.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
